\name{MCMCArmadillo-package}
\alias{MCMCArmadillo-package}
\alias{MCMCArmadillo}
\docType{package}
\title{
\packageTitle{MCMCArmadillo}
}
\description{
\packageDescription{MCMCArmadillo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MCMCArmadillo}
\packageIndices{MCMCArmadillo}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MCMCArmadillo}

Maintainer: \packageMaintainer{MCMCArmadillo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
